fix: conda environment file
- tell conda to look for mne in the conda-forge channel

Otherwise, (by default) it only looks in the "defaults" channell which does not have mne.
I specified the channel at the package level instead of the environment level because in my expereinece it leads to more reproducible environments.
Of course, we are not specifying package versions so any actual reproducibility is out of the question.
Still, at least having most of the packages from the "defaults" channels instead of an unpredictable mix of "defaults" and "conda-forge" is already something.

- specify the Python version - 3.7

Without this, conda took forever to solve the environment.
The reason for choosing 3.7 is simply that 3.8 feels too new for me at the moment.